SLVS-1564 Move IConfigScopeUpdater in the ConfigurationScope folder
