+++ src/Core/ConfigurationScope/IConfigScopeUpdater.cs
@@ -18,7 +18,9 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-namespace SonarLint.VisualStudio.Core.Binding;
+using SonarLint.VisualStudio.Core.Binding;
+
+namespace SonarLint.VisualStudio.Core.ConfigurationScope;
 
 /// <summary>
 /// Updates current active configuration scope of SLCore
